Make a light theme less bright.
The colours have been changed for softer ones. The commit covers most of the main Infirma colours for the light theme and separately code blocks. The colours of the titles in the footer have also been fixed (they were white in the light theme, making them invisible). Makes text in codeblocks `--ifm-color-primary-darker`.