Add shadows for better visibility of code blocks in light theme.
Added shadows to code blocks, menu details, and alert blocks. The main goal was to improve the visibility of code blocks in the light theme. Shadows were added to the other elements to maintain a unified style. Shadows were added to the dark theme for the same reason, also the lack of shadows in the dark theme caused blocks to distort when switching themes.